+++ tests/test_service.py
@@ -84,11 +84,6 @@ from control_plane.contracts.preview_generation_record import (
 )
 from control_plane.contracts.preview_inventory_scan_record import PreviewInventoryScanRecord
 from control_plane.contracts.preview_lifecycle_plan_record import PreviewLifecyclePlanRecord
-from control_plane.contracts.preview_pr_feedback_notifications import (
-    PreviewPrFeedbackNotificationAttemptRecord,
-    PreviewPrFeedbackNotificationDestination,
-    PreviewPrFeedbackNotificationPolicyRecord,
-)
 from control_plane.contracts.preview_record import PreviewRecord
 from control_plane.contracts.product_profile_record import LaunchplaneProductProfileRecord
 from control_plane.contracts.promotion_record import (
@@ -23932,25 +23927,9 @@ class LaunchplaneServiceTests(unittest.TestCase):
         self.assertEqual(payload["error"]["code"], "authorization_denied")
         self.assertEqual(payload["authz"]["request"]["product"], "verireel")
 
-    def test_preview_pr_feedback_endpoint_records_skipped_delivery_without_token(self) -> None:
+    def test_preview_pr_feedback_endpoint_legacy_wsgi_route_is_retired(self) -> None:
         with TemporaryDirectory() as temporary_directory_name:
             root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
             app = create_launchplane_service_app(
                 state_dir=root / "state",
                 verifier=_StubVerifier(
@@ -23961,7 +23940,7 @@ class LaunchplaneServiceTests(unittest.TestCase):
                         event_name="pull_request",
                     )
                 ),
-                authz_policy=policy,
+                authz_policy=LaunchplaneAuthzPolicy(),
                 control_plane_root_path=root,
             )
 
@@ -23978,1139 +23957,11 @@ class LaunchplaneServiceTests(unittest.TestCase):
                     "anchor_pr_number": 42,
                     "anchor_pr_url": "https://github.com/every/verireel/pull/42",
                     "status": "ready",
-                    "preview_url": "https://pr-42.preview.example",
-                    "immutable_image_reference": "ghcr.io/every/verireel:pr-42-a1b2c3d4",
-                    "refresh_image_reference": "ghcr.io/every/verireel:preview-pr-42",
-                    "revision": "a1b2c3d4",
-                    "run_url": "https://github.com/every/verireel/actions/runs/123",
                 },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="verireel-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(
-            payload["records"]["preview_pr_feedback_id"], feedback_records[0].feedback_id
-        )
-        self.assertEqual(payload["result"]["delivery_status"], "skipped")
-        self.assertIn("Launchplane preview is ready", payload["result"]["comment_markdown"])
-        self.assertIn("GITHUB_TOKEN", feedback_records[0].error_message)
-        self.assertEqual(feedback_records[0].anchor_pr_number, 42)
-
-    def test_preview_pr_feedback_skipped_delivery_dispatches_discord_notification(
-        self,
-    ) -> None:
-        sent_payloads: list[tuple[str, dict[str, object]]] = []
-
-        def send_discord(webhook_url: str, payload: dict[str, object]) -> None:
-            sent_payloads.append((webhook_url, payload))
-
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {control_plane_secrets.LAUNCHPLANE_SECRET_MASTER_KEY_ENV_VAR: "test-master-key"},
-                clear=True,
-            ),
-        ):
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-                database_url=database_url,
-                preview_pr_feedback_discord_sender=send_discord,
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                secret_result = control_plane_secrets.write_secret_value(
-                    record_store=store,
-                    scope="context_instance",
-                    integration="preview-pr-feedback-notifications",
-                    name="discord webhook",
-                    plaintext_value="https://discord.com/api/webhooks/test/webhook",
-                    binding_key="DISCORD_WEBHOOK",
-                    context_name="launchplane",
-                    instance_name="preview-feedback",
-                    actor="test",
-                    source_label="test",
-                )
-                store.write_preview_pr_feedback_notification_policy_record(
-                    PreviewPrFeedbackNotificationPolicyRecord(
-                        policy_id="preview-pr-feedback-notification-discord",
-                        product="verireel",
-                        context="verireel-testing",
-                        repository="every/verireel",
-                        status="enabled",
-                        created_at="2026-06-15T16:30:00Z",
-                        updated_at="2026-06-15T16:30:00Z",
-                        source="test",
-                        destinations=(
-                            PreviewPrFeedbackNotificationDestination(
-                                destination_id="discord",
-                                kind="discord",
-                                discord_webhook_secret=str(secret_result["secret_id"]),
-                            ),
-                        ),
-                    )
-                )
-            finally:
-                store.close()
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "source": "preview-control-plane",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 42,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/42",
-                    "status": "ready",
-                    "preview_url": "https://pr-42.preview.example",
-                    "run_url": "https://github.com/every/verireel/actions/runs/123",
-                },
-                headers={"Idempotency-Key": "preview-pr-feedback-notify-skipped"},
-            )
-            feedback_id = str(payload["records"]["preview_pr_feedback_id"])
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                attempts = store.list_preview_pr_feedback_notification_attempt_records(
-                    feedback_id=feedback_id,
-                    event="delivery_skipped",
-                )
-            finally:
-                store.close()
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["result"]["delivery_status"], "skipped")
-        self.assertEqual(len(payload["result"]["notifications"]), 1)
-        self.assertEqual(payload["result"]["notifications"][0]["delivery_status"], "delivered")
-        self.assertEqual(len(sent_payloads), 1)
-        self.assertEqual(sent_payloads[0][0], "https://discord.com/api/webhooks/test/webhook")
-        self.assertEqual(attempts[0].delivery_status, "delivered")
-        self.assertEqual(attempts[0].action, "posted_discord")
-
-    def test_preview_pr_feedback_failed_delivery_dispatches_discord_notification(
-        self,
-    ) -> None:
-        sent_payloads: list[tuple[str, dict[str, object]]] = []
-
-        def send_discord(webhook_url: str, payload: dict[str, object]) -> None:
-            sent_payloads.append((webhook_url, payload))
-
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {control_plane_secrets.LAUNCHPLANE_SECRET_MASTER_KEY_ENV_VAR: "test-master-key"},
-                clear=True,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.resolve_launchplane_github_token",
-                return_value="github-token",
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.find_github_issue_comment_by_marker",
-                return_value=None,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.create_github_issue_comment",
-                side_effect=ClickException("GitHub comments are unavailable."),
-            ),
-        ):
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/43/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/43/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-                database_url=database_url,
-                preview_pr_feedback_discord_sender=send_discord,
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                secret_result = control_plane_secrets.write_secret_value(
-                    record_store=store,
-                    scope="context_instance",
-                    integration="preview-pr-feedback-notifications",
-                    name="discord webhook",
-                    plaintext_value="https://discord.com/api/webhooks/test/webhook",
-                    binding_key="DISCORD_WEBHOOK",
-                    context_name="launchplane",
-                    instance_name="preview-feedback",
-                    actor="test",
-                    source_label="test",
-                )
-                store.write_preview_pr_feedback_notification_policy_record(
-                    PreviewPrFeedbackNotificationPolicyRecord(
-                        policy_id="preview-pr-feedback-notification-discord",
-                        product="verireel",
-                        context="verireel-testing",
-                        repository="every/verireel",
-                        status="enabled",
-                        created_at="2026-06-15T16:30:00Z",
-                        updated_at="2026-06-15T16:30:00Z",
-                        source="test",
-                        destinations=(
-                            PreviewPrFeedbackNotificationDestination(
-                                destination_id="discord",
-                                kind="discord",
-                                discord_webhook_secret=str(secret_result["secret_id"]),
-                            ),
-                        ),
-                    )
-                )
-            finally:
-                store.close()
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "source": "preview-control-plane",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 43,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/43",
-                    "status": "ready",
-                    "preview_url": "https://pr-43.preview.example",
-                },
-                headers={"Idempotency-Key": "preview-pr-feedback-notify-failed"},
-            )
-            feedback_id = str(payload["records"]["preview_pr_feedback_id"])
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                attempts = store.list_preview_pr_feedback_notification_attempt_records(
-                    feedback_id=feedback_id,
-                    event="delivery_failed",
-                )
-            finally:
-                store.close()
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["result"]["delivery_status"], "failed")
-        self.assertIn("GitHub comments are unavailable", payload["result"]["error_message"])
-        self.assertEqual(len(sent_payloads), 1)
-        self.assertEqual(attempts[0].delivery_status, "delivered")
-        self.assertEqual(attempts[0].action, "posted_discord")
-
-    def test_preview_pr_feedback_notification_attempt_read_route_is_retired_from_legacy_wsgi_app(
-        self,
-    ) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(_every_code_worker_identity()),
-                authz_policy=_every_code_worker_policy(
-                    extra_actions=("preview_pr_feedback_notification_attempt.read",)
-                ),
-                control_plane_root_path=root,
-                database_url=database_url,
-            )
-            attempt_record = PreviewPrFeedbackNotificationAttemptRecord(
-                attempt_id="preview-pr-feedback-notification-feedback-1-delivery-skipped",
-                feedback_id="feedback-1",
-                event="delivery_skipped",
-                policy_id="preview-pr-feedback-notification-discord",
-                destination_id="discord",
-                destination_kind="discord",
-                delivery_status="delivered",
-                attempted_at="2026-06-15T17:35:00Z",
-                action="posted_discord",
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                store.write_preview_pr_feedback_notification_attempt_record(attempt_record)
-            finally:
-                store.close()
-
-            status_code, payload = _invoke_app(
-                app,
-                method="GET",
-                path="/v1/previews/pr-feedback/notification-attempts",
-                query_string="feedback_id=feedback-1&event=delivery_skipped",
             )
 
         self.assertEqual(status_code, 404)
         self.assertEqual(payload["error"]["code"], "not_found")
-
-    def test_preview_pr_feedback_notification_attempt_read_authz_moved_to_fastapi(
-        self,
-    ) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(_every_code_worker_identity()),
-                authz_policy=_every_code_worker_policy(),
-                control_plane_root_path=root,
-                database_url=database_url,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="GET",
-                path="/v1/previews/pr-feedback/notification-attempts",
-            )
-
-        self.assertEqual(status_code, 404)
-        self.assertEqual(payload["error"]["code"], "not_found")
-
-    def test_preview_pr_feedback_notification_not_duplicated_on_idempotent_retry(
-        self,
-    ) -> None:
-        sent_payloads: list[tuple[str, dict[str, object]]] = []
-
-        def send_discord(webhook_url: str, payload: dict[str, object]) -> None:
-            sent_payloads.append((webhook_url, payload))
-
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {control_plane_secrets.LAUNCHPLANE_SECRET_MASTER_KEY_ENV_VAR: "test-master-key"},
-                clear=True,
-            ),
-        ):
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/44/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/44/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-                database_url=database_url,
-                preview_pr_feedback_discord_sender=send_discord,
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                secret_result = control_plane_secrets.write_secret_value(
-                    record_store=store,
-                    scope="context_instance",
-                    integration="preview-pr-feedback-notifications",
-                    name="discord webhook",
-                    plaintext_value="https://discord.com/api/webhooks/test/webhook",
-                    binding_key="DISCORD_WEBHOOK",
-                    context_name="launchplane",
-                    instance_name="preview-feedback",
-                    actor="test",
-                    source_label="test",
-                )
-                store.write_preview_pr_feedback_notification_policy_record(
-                    PreviewPrFeedbackNotificationPolicyRecord(
-                        policy_id="preview-pr-feedback-notification-discord",
-                        product="verireel",
-                        context="verireel-testing",
-                        repository="every/verireel",
-                        status="enabled",
-                        created_at="2026-06-15T16:30:00Z",
-                        updated_at="2026-06-15T16:30:00Z",
-                        source="test",
-                        destinations=(
-                            PreviewPrFeedbackNotificationDestination(
-                                destination_id="discord",
-                                kind="discord",
-                                discord_webhook_secret=str(secret_result["secret_id"]),
-                            ),
-                        ),
-                    )
-                )
-            finally:
-                store.close()
-
-            request_payload = {
-                "product": "verireel",
-                "context": "verireel-testing",
-                "source": "preview-control-plane",
-                "repository": "every/verireel",
-                "anchor_repo": "verireel",
-                "anchor_pr_number": 44,
-                "anchor_pr_url": "https://github.com/every/verireel/pull/44",
-                "status": "ready",
-                "preview_url": "https://pr-44.preview.example",
-            }
-            headers = {"Idempotency-Key": "preview-pr-feedback-notify-retry"}
-            first_status, first_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload=request_payload,
-                headers=headers,
-            )
-            second_status, second_payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload=request_payload,
-                headers=headers,
-            )
-            feedback_id = str(first_payload["records"]["preview_pr_feedback_id"])
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                attempts = store.list_preview_pr_feedback_notification_attempt_records(
-                    feedback_id=feedback_id,
-                    event="delivery_skipped",
-                )
-            finally:
-                store.close()
-
-        self.assertEqual(first_status, 202)
-        self.assertEqual(second_status, 202)
-        self.assertEqual(second_payload["replayed"], True)
-        self.assertEqual(len(sent_payloads), 1)
-        self.assertEqual(len(attempts), 1)
-
-    def test_preview_pr_feedback_successful_delivery_does_not_dispatch_discord(
-        self,
-    ) -> None:
-        sent_payloads: list[tuple[str, dict[str, object]]] = []
-
-        def send_discord(webhook_url: str, payload: dict[str, object]) -> None:
-            sent_payloads.append((webhook_url, payload))
-
-        with (
-            TemporaryDirectory() as temporary_directory_name,
-            patch.dict(
-                os.environ,
-                {control_plane_secrets.LAUNCHPLANE_SECRET_MASTER_KEY_ENV_VAR: "test-master-key"},
-                clear=True,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.resolve_launchplane_github_token",
-                return_value="github-token",
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.find_github_issue_comment_by_marker",
-                return_value=None,
-            ),
-            patch(
-                "control_plane.workflows.preview_pr_feedback.create_github_issue_comment",
-                return_value={
-                    "id": 123,
-                    "html_url": "https://github.com/every/verireel/pull/45#issuecomment-123",
-                },
-            ),
-        ):
-            root = Path(temporary_directory_name)
-            database_url = _sqlite_database_url(root / "launchplane.sqlite3")
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/45/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/45/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-                database_url=database_url,
-                preview_pr_feedback_discord_sender=send_discord,
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                store.write_preview_pr_feedback_notification_policy_record(
-                    PreviewPrFeedbackNotificationPolicyRecord(
-                        policy_id="preview-pr-feedback-notification-discord",
-                        product="verireel",
-                        context="verireel-testing",
-                        repository="every/verireel",
-                        status="enabled",
-                        created_at="2026-06-15T16:30:00Z",
-                        updated_at="2026-06-15T16:30:00Z",
-                        source="test",
-                        destinations=(
-                            PreviewPrFeedbackNotificationDestination(
-                                destination_id="discord",
-                                kind="discord",
-                                discord_webhook_secret="secret-id",
-                            ),
-                        ),
-                    )
-                )
-            finally:
-                store.close()
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "source": "preview-control-plane",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 45,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/45",
-                    "status": "pending",
-                },
-                headers={"Idempotency-Key": "preview-pr-feedback-notify-success"},
-            )
-            store = PostgresRecordStore(database_url=database_url)
-            try:
-                attempts = store.list_preview_pr_feedback_notification_attempt_records()
-            finally:
-                store.close()
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["result"]["delivery_status"], "delivered")
-        self.assertEqual(sent_payloads, [])
-        self.assertEqual(attempts, ())
-
-    def test_preview_pr_feedback_endpoint_dry_run_authorizes_without_writing(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "source": "preview-control-plane-smoke",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 42,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/42",
-                    "status": "unsupported",
-                    "failure_summary": "route smoke only",
-                    "dry_run": True,
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="verireel-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(payload["result"]["dry_run"], True)
-        self.assertEqual(payload["result"]["preview_pr_feedback"], "authorized")
-        self.assertEqual(payload["result"]["status"], "unsupported")
-        self.assertEqual(payload["records"], {})
-        self.assertEqual(feedback_records, ())
-
-    def test_preview_pr_feedback_endpoint_records_unsupported_notice(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-fork-notice.yml@refs/heads/main"
-                            ],
-                            "event_names": ["pull_request_target"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-fork-notice.yml@refs/heads/main"
-                        ),
-                        event_name="pull_request_target",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "source": "preview-fork-notice",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 43,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/43",
-                    "status": "unsupported",
-                    "run_url": "https://github.com/every/verireel/actions/runs/124",
-                    "failure_summary": "Preview automation is unavailable for forked pull requests.",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="verireel-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].status, "unsupported")
-        self.assertIn("preview automation is unavailable", payload["result"]["comment_markdown"])
-        self.assertIn("protected preview provisioning path", payload["result"]["comment_markdown"])
-        self.assertEqual(payload["result"]["delivery_status"], "skipped")
-
-    def test_preview_pr_feedback_endpoint_accepts_generic_refresh_grant(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/sellyouroutboard",
-                            "workflow_refs": [
-                                "cbusillo/sellyouroutboard/.github/workflows/preview-control-plane.yml@refs/pull/19/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["sellyouroutboard"],
-                            "contexts": ["sellyouroutboard-testing"],
-                            "actions": ["preview_refresh.execute"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/sellyouroutboard",
-                        workflow_ref=(
-                            "cbusillo/sellyouroutboard/.github/workflows/preview-control-plane.yml"
-                            "@refs/pull/19/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "sellyouroutboard",
-                    "context": "sellyouroutboard-testing",
-                    "source": "preview-control-plane",
-                    "repository": "cbusillo/sellyouroutboard",
-                    "anchor_repo": "sellyouroutboard",
-                    "anchor_pr_number": 19,
-                    "anchor_pr_url": "https://github.com/cbusillo/sellyouroutboard/pull/19",
-                    "status": "ready",
-                    "preview_url": "https://pr-19.syo-preview.shinycomputers.com",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="sellyouroutboard-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].status, "ready")
-
-    def test_preview_pr_feedback_endpoint_accepts_pending_refresh_grant(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/sellyouroutboard",
-                            "workflow_refs": [
-                                "cbusillo/sellyouroutboard/.github/workflows/preview-fork-notice.yml@refs/heads/main"
-                            ],
-                            "event_names": ["pull_request_target"],
-                            "products": ["sellyouroutboard"],
-                            "contexts": ["sellyouroutboard-testing"],
-                            "actions": ["preview_refresh.execute"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/sellyouroutboard",
-                        workflow_ref=(
-                            "cbusillo/sellyouroutboard/.github/workflows/preview-fork-notice.yml"
-                            "@refs/heads/main"
-                        ),
-                        event_name="pull_request_target",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "sellyouroutboard",
-                    "context": "sellyouroutboard-testing",
-                    "source": "preview-fork-notice",
-                    "repository": "cbusillo/sellyouroutboard",
-                    "anchor_repo": "sellyouroutboard",
-                    "anchor_pr_number": 19,
-                    "anchor_pr_url": "https://github.com/cbusillo/sellyouroutboard/pull/19",
-                    "status": "pending",
-                    "run_url": "https://github.com/cbusillo/sellyouroutboard/actions/runs/123",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="sellyouroutboard-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].status, "pending")
-        self.assertIn("preview is waiting", payload["result"]["comment_markdown"])
-        self.assertIn("prerequisites are still in flight", payload["result"]["comment_markdown"])
-
-    def test_preview_pr_feedback_endpoint_accepts_generic_destroy_grant(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/sellyouroutboard",
-                            "workflow_refs": [
-                                "cbusillo/sellyouroutboard/.github/workflows/preview-cleanup.yml@refs/heads/main"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["sellyouroutboard"],
-                            "contexts": ["sellyouroutboard-testing"],
-                            "actions": ["preview_destroy.execute"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/sellyouroutboard",
-                        workflow_ref=(
-                            "cbusillo/sellyouroutboard/.github/workflows/preview-cleanup.yml"
-                            "@refs/heads/main"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "sellyouroutboard",
-                    "context": "sellyouroutboard-testing",
-                    "source": "preview-cleanup",
-                    "repository": "cbusillo/sellyouroutboard",
-                    "anchor_repo": "sellyouroutboard",
-                    "anchor_pr_number": 19,
-                    "anchor_pr_url": "https://github.com/cbusillo/sellyouroutboard/pull/19",
-                    "status": "destroyed",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="sellyouroutboard-testing")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].status, "destroyed")
-
-    def test_preview_pr_feedback_endpoint_keeps_unsupported_explicit(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/sellyouroutboard",
-                            "workflow_refs": [
-                                "cbusillo/sellyouroutboard/.github/workflows/preview-control-plane.yml@refs/pull/19/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["sellyouroutboard"],
-                            "contexts": ["sellyouroutboard-testing"],
-                            "actions": ["preview_refresh.execute"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/sellyouroutboard",
-                        workflow_ref=(
-                            "cbusillo/sellyouroutboard/.github/workflows/preview-control-plane.yml"
-                            "@refs/pull/19/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "sellyouroutboard",
-                    "context": "sellyouroutboard-testing",
-                    "source": "preview-control-plane",
-                    "repository": "cbusillo/sellyouroutboard",
-                    "anchor_repo": "sellyouroutboard",
-                    "anchor_pr_number": 19,
-                    "anchor_pr_url": "https://github.com/cbusillo/sellyouroutboard/pull/19",
-                    "status": "unsupported",
-                },
-            )
-
-        self.assertEqual(status_code, 403)
-        self.assertEqual(payload["error"]["code"], "authorization_denied")
-
-    def test_preview_pr_feedback_endpoint_accepts_odoo_cm_unsupported_feedback_grant(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/odoo-tenant-cm",
-                            "workflow_refs": [
-                                "cbusillo/odoo-tenant-cm/.github/workflows/odoo-preview.yml@refs/pull/27/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["odoo-tenant-cm"],
-                            "contexts": ["cm"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/odoo-tenant-cm",
-                        workflow_ref=(
-                            "cbusillo/odoo-tenant-cm/.github/workflows/odoo-preview.yml"
-                            "@refs/pull/27/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "odoo-tenant-cm",
-                    "context": "cm",
-                    "source": "odoo-preview-unsupported",
-                    "repository": "cbusillo/odoo-tenant-cm",
-                    "anchor_repo": "odoo-tenant-cm",
-                    "anchor_pr_number": 27,
-                    "anchor_pr_url": "https://github.com/cbusillo/odoo-tenant-cm/pull/27",
-                    "status": "unsupported",
-                    "run_url": "https://github.com/cbusillo/odoo-tenant-cm/actions/runs/123",
-                    "failure_summary": "Preview automation is unavailable for forked pull requests.",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="cm")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].product, "odoo-tenant-cm")
-        self.assertEqual(feedback_records[0].status, "unsupported")
-        self.assertIn("preview automation is unavailable", payload["result"]["comment_markdown"])
-
-    def test_preview_pr_feedback_endpoint_accepts_odoo_cm_pull_request_target_feedback_grant(
-        self,
-    ) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "cbusillo/odoo-tenant-cm",
-                            "workflow_refs": [
-                                "cbusillo/odoo-tenant-cm/.github/workflows/odoo-preview.yml@refs/heads/main"
-                            ],
-                            "event_names": ["pull_request_target"],
-                            "products": ["odoo-tenant-cm"],
-                            "contexts": ["cm"],
-                            "actions": ["preview_pr_feedback.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        repository="cbusillo/odoo-tenant-cm",
-                        workflow_ref=(
-                            "cbusillo/odoo-tenant-cm/.github/workflows/odoo-preview.yml"
-                            "@refs/heads/main"
-                        ),
-                        event_name="pull_request_target",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "odoo-tenant-cm",
-                    "context": "cm",
-                    "source": "odoo-preview-unsupported",
-                    "repository": "cbusillo/odoo-tenant-cm",
-                    "anchor_repo": "odoo-tenant-cm",
-                    "anchor_pr_number": 27,
-                    "anchor_pr_url": "https://github.com/cbusillo/odoo-tenant-cm/pull/27",
-                    "status": "unsupported",
-                    "run_url": "https://github.com/cbusillo/odoo-tenant-cm/actions/runs/123",
-                    "failure_summary": "Preview automation is unavailable for forked pull requests.",
-                },
-            )
-
-            feedback_records = FilesystemRecordStore(
-                state_dir=root / "state"
-            ).list_preview_pr_feedback_records(context_name="cm")
-
-        self.assertEqual(status_code, 202)
-        self.assertEqual(payload["status"], "accepted")
-        self.assertEqual(feedback_records[0].status, "unsupported")
-
-    def test_preview_pr_feedback_endpoint_rejects_unauthorized_workflow(self) -> None:
-        with TemporaryDirectory() as temporary_directory_name:
-            root = Path(temporary_directory_name)
-            policy = LaunchplaneAuthzPolicy.model_validate(
-                {
-                    "github_actions": [
-                        {
-                            "repository": "every/verireel",
-                            "workflow_refs": [
-                                "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                            ],
-                            "event_names": ["pull_request"],
-                            "products": ["verireel"],
-                            "contexts": ["verireel-testing"],
-                            "actions": ["preview_generation.write"],
-                        }
-                    ]
-                }
-            )
-            app = create_launchplane_service_app(
-                state_dir=root / "state",
-                verifier=_StubVerifier(
-                    _identity(
-                        workflow_ref=(
-                            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge"
-                        ),
-                        event_name="pull_request",
-                    )
-                ),
-                authz_policy=policy,
-                control_plane_root_path=root,
-            )
-
-            status_code, payload = _invoke_app(
-                app,
-                method="POST",
-                path="/v1/previews/pr-feedback",
-                payload={
-                    "product": "verireel",
-                    "context": "verireel-testing",
-                    "repository": "every/verireel",
-                    "anchor_repo": "verireel",
-                    "anchor_pr_number": 42,
-                    "anchor_pr_url": "https://github.com/every/verireel/pull/42",
-                    "status": "ready",
-                },
-            )
-
-        self.assertEqual(status_code, 403)
-        self.assertEqual(payload["error"]["code"], "authorization_denied")
-        self.assertEqual(payload["authz"]["identity"]["repository"], "every/verireel")
-        self.assertEqual(
-            payload["authz"]["identity"]["workflow_ref"],
-            "every/verireel/.github/workflows/preview-control-plane.yml@refs/pull/42/merge",
-        )
-        self.assertEqual(payload["authz"]["identity"]["event_name"], "pull_request")
-        self.assertEqual(payload["authz"]["request"]["action"], "preview_pr_feedback.write")
-        self.assertEqual(payload["authz"]["request"]["product"], "verireel")
-        self.assertEqual(payload["authz"]["request"]["context"], "verireel-testing")
-        self.assertEqual(payload["authz"]["agent_audit"]["decision"], "denied")
-        self.assertEqual(payload["authz"]["agent_audit"]["reason_code"], "authorization_denied")
-        self.assertEqual(payload["authz"]["agent_audit"]["subject"]["action_safety"], "safe_write")
-        self.assertEqual(payload["authz"]["agent_audit"]["source_kind"], "authz_policy")
-        self.assertIn("policy_sha256", payload["authz"])
-        self.assertIn("policy_source", payload["authz"])
 
     def test_preview_lifecycle_cleanup_endpoint_records_report_only_cleanup(self) -> None:
         with TemporaryDirectory() as temporary_directory_name:
